+++ tests/testthat/test-output.R
@@ -1,26 +1,21 @@
 context("Output")
 
-test_that("different outputs produce identical code", {
-  simple_plot <- function(path) {
-    svglite(path)
-    plot(1:5)
-    invisible(dev.off())
-  }
-
+test_that("different string and file output produce identical svg", {
   ## 1. Write to a file
   f1 <- tempfile()
-  simple_plot(f1)
+  svglite(f1)
+  plot(1:5)
+  dev.off()
+
   out1 <- readLines(f1)
 
-  ## 2. Direct to R terminal
-  out2 <- capture.output(simple_plot(":terminal:"))
+  ## 2. Write to a string stream
+  s <- svgstring()
+  plot(1:5)
+  dev.off()
+  out2 <- strsplit(s(), "\n")[[1]]
 
-  ## 3. Write to a string stream
-  simple_plot(":string:")
-  out3 <- strsplit(stringSVG(), "\n")[[1]]
-
-  # expect_equal(out1, out2)
-  expect_equal(out1, out3)
+  expect_equal(out1, out2)
 })
 
 test_that("intermediate outputs are always valid svg", {