+++ NEWS.md
@@ -5,9 +5,7 @@
 
 * Intermediate outputs are always valid SVG (#53).
 
-* `svglite` accepts special destinations `:terminal:` and `:string:` to print
-  svg to the terminal or return as a string (accessed with `stringSVG()`) 
-  respectively (#40, @yixuan).
+* New `svgstring()` returns plot as a string (#40, @yixuan).
 
 * Use raster test compatible with older versions of R.